Fix casing of SideMenuComponent and MobileSideMenuComponent
