+++ src/layouts/header/HeaderComponent.tsx
@@ -7,7 +7,7 @@ import SvgIcon from '@/components/SvgIcon'
 
 const HeaderComponent : React.FC = () => {
   return (
-    <div className='flex header-height w-full items-center bg-white border-b border-[#A6B4DA]'>
+    <div className='header-height flex w-full items-center bg-white border-b border-[#A6B4DA]'>
       <div className='sidebar-width relative h-full hidden tablet:flex justify-center items-center border-r border-[#A6B4DA] px-4'>
       <SvgIcon name={"hotelLogo"} width={151} height={46}/>
       <CollapseSideMenu />
