+++ src/layouts/header/right-panel/RightPanelComponent.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ProfileComponent from './ProfileComponent';
 import IconBadgeContainer from './IconBadgeContainer';
-import { BadgeType } from '../../../../types/enum/BadgeType';
-import MobileSidemenuComponent from '@/layouts/sidemenu/MobileSidemenuComponent';
+import { BadgeType } from '../../../types/enum/BadgeType';
+import MobileSideMenuComponent from '@/layouts/sidemenu/MobileSideMenuComponent';
 
 const RightPanelComponent : React.FC  = () => {
   return (
@@ -12,7 +12,7 @@ const RightPanelComponent : React.FC  = () => {
       <div className='hidden tablet:flex'>
       <ProfileComponent />
       </div>
-        <div className='flex tablet:hidden'><MobileSidemenuComponent /></div>
+        <div className='flex tablet:hidden'><MobileSideMenuComponent /></div>
     </div>
   )
 }
