+++ src/types/enum/BadgeType.ts
